+++ src/main/java/org/monarchinitiative/hpoannotqc/exception/HpoAnnotationFileException.java
@@ -4,10 +4,10 @@ import org.monarchinitiative.phenol.base.PhenolException;
 
 /**
  * Exceptions of this class are thrown if there are quality control issues with the
- * generation of a SmallFile or SmallFileEntry (single line of an HPO annotation file)
+ * generation of a HpoAnnotationFile or HpoAnnotationFileEntry (single line of an HPO annotation file)
  * @author Peter Robinson
  */
-public class SmallFileException extends PhenolException {
-    public SmallFileException(String msg) {super(msg);}
+public class HpoAnnotationFileException extends PhenolException {
+    public HpoAnnotationFileException(String msg) {super(msg);}
 
 }